feat: separate runtime domain from persistence
