+++ runtime-kotlin/src/main/kotlin/skillbill/learnings/LearningsRuntime.kt
@@ -1,76 +1,68 @@
 package skillbill.learnings
 
-import skillbill.contracts.JsonSupport
-import skillbill.review.LearningRecord
-import skillbill.review.ReviewRuntime
-import java.sql.Connection
+data class RejectedLearningSourceOutcome(
+  val eventType: String,
+  val note: String,
+)
+
+data class LearningSourceReference(
+  val reviewRunId: String,
+  val findingId: String,
+)
 
 data class LearningSourceValidation(
   val reviewRunId: String,
   val findingId: String,
-  val rejectedOutcome: Pair<String, String>,
+  val rejectedOutcome: RejectedLearningSourceOutcome,
 )
 
 object LearningsRuntime {
   val learningStatuses: List<String> = listOf("active", "disabled")
-  private val rejectedFindingOutcomeTypes = listOf("fix_rejected", "false_positive")
-  private const val REJECTED_OUTCOME_FIRST_PARAM_INDEX: Int = 3
+  val rejectedFindingOutcomeTypes: List<String> = listOf("fix_rejected", "false_positive")
 
   fun validateLearningScope(scope: LearningScope, scopeKey: String): Pair<LearningScope, String> =
     scope to scope.normalizeScopeKey(scopeKey)
 
-  fun validateLearningSource(
-    connection: Connection,
-    sourceReviewRunId: String?,
-    sourceFindingId: String?,
-  ): LearningSourceValidation {
+  fun validateLearningSourceReference(sourceReviewRunId: String?, sourceFindingId: String?): LearningSourceReference {
     require(!sourceReviewRunId.isNullOrBlank() && !sourceFindingId.isNullOrBlank()) {
       "Learnings must be derived from a rejected review finding. Provide both --from-run and --from-finding."
     }
-    require(ReviewRuntime.findingExists(connection, sourceReviewRunId, sourceFindingId)) {
-      "Unknown learning source '$sourceReviewRunId:$sourceFindingId'. Import the review and finding first."
+    return LearningSourceReference(
+      reviewRunId = sourceReviewRunId.trim(),
+      findingId = sourceFindingId.trim(),
+    )
+  }
+
+  fun validateLearningSource(
+    sourceReviewRunId: String?,
+    sourceFindingId: String?,
+    sourceFindingExists: Boolean,
+    latestRejectedOutcome: RejectedLearningSourceOutcome?,
+  ): LearningSourceValidation = validateLearningSource(
+    sourceReference = validateLearningSourceReference(sourceReviewRunId, sourceFindingId),
+    sourceFindingExists = sourceFindingExists,
+    latestRejectedOutcome = latestRejectedOutcome,
+  )
+
+  fun validateLearningSource(
+    sourceReference: LearningSourceReference,
+    sourceFindingExists: Boolean,
+    latestRejectedOutcome: RejectedLearningSourceOutcome?,
+  ): LearningSourceValidation {
+    require(sourceFindingExists) {
+      "Unknown learning source '${sourceReference.reviewRunId}:${sourceReference.findingId}'. " +
+        "Import the review and finding first."
     }
-    val rejectedOutcome = fetchLatestRejectedOutcome(connection, sourceReviewRunId, sourceFindingId)
-    require(rejectedOutcome != null) {
-      "Finding '$sourceFindingId' in run '$sourceReviewRunId' has no rejected outcome. " +
+    require(latestRejectedOutcome != null) {
+      "Finding '${sourceReference.findingId}' in run '${sourceReference.reviewRunId}' has no rejected outcome. " +
         "Learnings can only be created from findings the user rejected " +
         "(fix_rejected or false_positive)."
     }
     return LearningSourceValidation(
-      reviewRunId = sourceReviewRunId,
-      findingId = sourceFindingId,
-      rejectedOutcome = rejectedOutcome,
+      reviewRunId = sourceReference.reviewRunId,
+      findingId = sourceReference.findingId,
+      rejectedOutcome = latestRejectedOutcome,
     )
-  }
-
-  fun fetchLatestRejectedOutcome(
-    connection: Connection,
-    reviewRunId: String,
-    findingId: String,
-  ): Pair<String, String>? {
-    val placeholders = rejectedFindingOutcomeTypes.joinToString(", ") { "?" }
-    return connection.prepareStatement(
-      """
-      SELECT event_type, note
-      FROM feedback_events
-      WHERE review_run_id = ? AND finding_id = ? AND event_type IN ($placeholders)
-      ORDER BY id DESC
-      LIMIT 1
-      """.trimIndent(),
-    ).use { statement ->
-      statement.setString(1, reviewRunId)
-      statement.setString(2, findingId)
-      rejectedFindingOutcomeTypes.forEachIndexed { index, value ->
-        statement.setString(index + REJECTED_OUTCOME_FIRST_PARAM_INDEX, value)
-      }
-      statement.executeQuery().use { resultSet ->
-        if (resultSet.next()) {
-          resultSet.getString("event_type") to resultSet.getString("note").orEmpty()
-        } else {
-          null
-        }
-      }
-    }
   }
 
   fun normalizeOptionalLookupValue(rawValue: String?, argumentName: String): String? {
@@ -82,103 +74,25 @@ object LearningsRuntime {
     return normalized
   }
 
-  fun resolveLearnings(
-    connection: Connection,
-    repoScopeKey: String?,
-    skillName: String?,
-  ): Triple<String?, String?, List<LearningRecord>> {
-    val normalizedRepoScopeKey = normalizeOptionalLookupValue(repoScopeKey, "--repo")
-    val normalizedSkillName = normalizeOptionalLookupValue(skillName, "--skill")
-    val scopeClauses = mutableListOf("scope = '${LearningScope.GLOBAL.wireName}'")
-    val parameters = mutableListOf<String>()
-    if (normalizedRepoScopeKey != null) {
-      scopeClauses += "(scope = '${LearningScope.REPO.wireName}' AND scope_key = ?)"
-      parameters += normalizedRepoScopeKey
+  fun validateLearningStatus(status: String): String {
+    require(status in learningStatuses) {
+      "Learning status must be one of ${learningStatuses.joinToString(", ")}."
     }
-    if (normalizedSkillName != null) {
-      scopeClauses += "(scope = '${LearningScope.SKILL.wireName}' AND scope_key = ?)"
-      parameters += normalizedSkillName
-    }
-
-    val rows =
-      connection.prepareStatement(
-        """
-        SELECT
-          id,
-          scope,
-          scope_key,
-          title,
-          rule_text,
-          rationale,
-          status,
-          source_review_run_id,
-          source_finding_id,
-          created_at,
-          updated_at
-        FROM learnings
-        WHERE status = 'active'
-          AND (${scopeClauses.joinToString(" OR ")})
-        ORDER BY
-          ${learningScopeOrderClause("scope")},
-          id
-        """.trimIndent(),
-      ).use { statement ->
-        parameters.forEachIndexed { index, parameter ->
-          statement.setString(index + 1, parameter)
-        }
-        statement.executeQuery().use { resultSet ->
-          buildList {
-            while (resultSet.next()) {
-              add(LearningStore.getLearning(connection, resultSet.getInt("id")))
-            }
-          }
-        }
-      }
-    return Triple(normalizedRepoScopeKey, normalizedSkillName, rows)
+    return status
   }
 
-  fun saveSessionLearnings(connection: Connection, reviewSessionId: String, learningsJson: String) {
-    connection.prepareStatement(
-      """
-      INSERT INTO session_learnings (review_session_id, learnings_json, updated_at)
-      VALUES (?, ?, CURRENT_TIMESTAMP)
-      ON CONFLICT(review_session_id) DO UPDATE SET
-        learnings_json = excluded.learnings_json,
-        updated_at = CURRENT_TIMESTAMP
-      """.trimIndent(),
-    ).use { statement ->
-      statement.setString(1, reviewSessionId)
-      statement.setString(2, learningsJson)
-      statement.executeUpdate()
-    }
+  fun validateLearningText(title: String, ruleText: String): Pair<String, String> {
+    val normalizedTitle = title.trim()
+    val normalizedRuleText = ruleText.trim()
+    require(normalizedTitle.isNotEmpty()) { "Learning title must not be empty." }
+    require(normalizedRuleText.isNotEmpty()) { "Learning rule text must not be empty." }
+    return normalizedTitle to normalizedRuleText
   }
 
-  fun fetchSessionLearnings(connection: Connection, reviewSessionId: String): Map<String, Any?>? =
-    connection.prepareStatement(
-      """
-      SELECT learnings_json
-      FROM session_learnings
-      WHERE review_session_id = ?
-      """.trimIndent(),
-    ).use { statement ->
-      statement.setString(1, reviewSessionId)
-      statement.executeQuery().use { resultSet ->
-        if (!resultSet.next()) {
-          return null
-        }
-        decodeSessionLearnings(resultSet.getString("learnings_json"))
-      }
+  fun effectiveRationale(rationale: String, rejectedOutcomeNote: String): String =
+    if (rationale.trim().isEmpty() && rejectedOutcomeNote.isNotBlank()) {
+      rejectedOutcomeNote.trim()
+    } else {
+      rationale.trim()
     }
-}
-
-private fun learningScopeOrderClause(columnName: String): String = buildString {
-  appendLine("CASE $columnName")
-  LearningScope.precedence.forEachIndexed { index, scope ->
-    appendLine("  WHEN '${scope.wireName}' THEN $index")
-  }
-  append("  ELSE ${LearningScope.precedence.size}\nEND")
-}
-
-private fun decodeSessionLearnings(rawJson: String): Map<String, Any?>? = JsonSupport.parseObjectOrNull(rawJson)?.let {
-  JsonSupport.anyToStringAnyMap(JsonSupport.jsonElementToValue(it))
 }
