+++ runtime-kotlin/src/main/kotlin/skillbill/ports/persistence/ReviewRepository.kt
@@ -1,5 +1,6 @@
 package skillbill.ports.persistence
 
+import skillbill.learnings.RejectedLearningSourceOutcome
 import skillbill.review.FeedbackRequest
 import skillbill.review.FeedbackTelemetryOptions
 import skillbill.review.ImportedReview
@@ -15,6 +16,10 @@ interface ReviewRepository {
   fun recordFeedback(request: FeedbackRequest, telemetryOptions: FeedbackTelemetryOptions): Map<String, Any?>?
 
   fun fetchNumberedFindings(runId: String): List<NumberedFinding>
+
+  fun findingExists(runId: String, findingId: String): Boolean
+
+  fun latestRejectedLearningSourceOutcome(runId: String, findingId: String): RejectedLearningSourceOutcome?
 
   fun reviewStatsPayload(runId: String?): Map<String, Any?>
