+++ runtime-kotlin/src/main/kotlin/skillbill/review/ReviewModels.kt
@@ -66,17 +66,3 @@ data class FindingOutcomeRow(
   val outcomeType: String,
   val note: String,
 )
-
-data class LearningRecord(
-  val id: Int,
-  val scope: String,
-  val scopeKey: String,
-  val title: String,
-  val ruleText: String,
-  val rationale: String,
-  val status: String,
-  val sourceReviewRunId: String?,
-  val sourceFindingId: String?,
-  val createdAt: String,
-  val updatedAt: String,
-)
