+++ runtime-kotlin/src/main/kotlin/skillbill/review/ReviewParser.kt
@@ -1,0 +1,29 @@
+package skillbill.review
+
+object ReviewParser {
+  fun parseReview(text: String): ImportedReview {
+    val reviewRunId =
+      requireMatch(
+        reviewRunIdPattern,
+        text,
+        "Review output is missing 'Review run ID: <review-run-id>'.",
+      )
+    val reviewSessionId =
+      requireMatch(
+        reviewSessionIdPattern,
+        text,
+        "Review output is missing 'Review session ID: <review-session-id>'.",
+      )
+    return ImportedReview(
+      reviewRunId = reviewRunId,
+      reviewSessionId = reviewSessionId,
+      rawText = text,
+      routedSkill = extractSummaryValue(text, "routed_skill"),
+      detectedScope = extractSummaryValue(text, "detected_scope"),
+      detectedStack = extractSummaryValue(text, "detected_stack"),
+      executionMode = extractSummaryValue(text, "execution_mode"),
+      specialistReviews = extractSpecialistReviews(text),
+      findings = parseReviewFindings(text),
+    )
+  }
+}
